delete the existing files while cloasing the browser

+++ icam-expert-backend/src/rest/queryDocuments/queryDocuments.controller.ts
@@ -2,6 +2,7 @@ import EnvVars from '@src/constants/EnvVars';
 import { NodeEnvs } from '@src/constants/misc';
 import { Request, Response } from 'express';
 import * as fs from 'fs';
+import { readFile, writeFile } from 'fs/promises';
 import multer, { Multer } from 'multer';
 import path from 'path';
 import { config } from '../../appConfig';
@@ -18,14 +19,13 @@ const audioFileTypes = ['.wav', '.ogg', '.mp3', '.flac', '.amr'];
 const videoFileTypes = ['.mp4', '.avi', '.mkv', '.mov', '.flv', '.wmv', '.webm', '.mpeg', '.mpg', '.3gp', '.ogv'];
 
 const docFolder = config.appSettings.uploadFolder;
+console.log('checking if uploads folder exists');
 if (!fs.existsSync(docFolder)) {
   console.log('Uploads folder does not exist, creating...');
   fs.mkdirSync(docFolder);
 }
 
 const saveFiles = (req: Request, res: Response, next: () => void) => {
-  console.log('checking if uploads folder exists');
-
   const storage = multer.diskStorage({
     destination: function (req, file, cb) {
       cb(null, docFolder);
@@ -33,14 +33,18 @@ const saveFiles = (req: Request, res: Response, next: () => void) => {
     filename: function (req, file, cb) {
       console.log('Checking if the file itself exists...');
 
-      if (fs.existsSync(path.join(docFolder, file.originalname))) {
-        // File exists. Add the suffix (1) to the filename and save the new one.
-        console.log('File exists. Adding an additional suffix and creating the new file...');
-        const parsedFile = path.parse(file.originalname);
-        cb(null, `${parsedFile.name}(1)${parsedFile.ext}`);
-      } else {
-        cb(null, file.originalname);
-      }
+      // TODO: #43 Reuse file -  always use the same filename (assumed the same file for now)
+      // if (fs.existsSync(path.join(docFolder, file.originalname))) {
+      //// File exists. Add the suffix (1) to the filename and save the new one.
+
+      //// TODO as part of #43, compare with the existing file and: save as new file (1) if different.
+
+      // console.log('File exists. Adding an additional suffix and creating the new file...');
+      // const parsedFile = path.parse(file.originalname);
+      // cb(null, `${parsedFile.name}(1)${parsedFile.ext}`);
+      // } else {
+      cb(null, file.originalname);
+      // }
     },
   });
 
@@ -120,28 +124,46 @@ const extractTextDocumentsFromFile = async (files: Express.Multer.File[]): Promi
 
       let outputText = '';
 
-      if (path.extname(originalFilePath) === '.docx') {
-        // Read docx file and get the plain text
-        outputText = await convertWordDocToText(originalFilePath);
-      } else if (viewerFileTypes.includes(path.extname(originalFilePath))) {
-        // Read the handwritten or any meaningful text directly
-        outputText = await convertHandwrittenFileToTextByAzure(originalFilePath);
-      } else if (audioFileTypes.includes(path.extname(originalFilePath))) {
-        console.log('AUDIO BEING PROCESSED: ', originalFilePath);
-        // Convert audio to text
-        outputText = await convertSpeechToText(originalFilePath);
+      // TODO Check if the output text exists for this file. e.g. myFile.pdf.text
+      // If it exists, skip the following conversion
+      // If not, go through the following conversions and then save the new text file myFile.pdf.text
 
-        console.log('AUDIO TEXT OUTPUT: ', outputText);
-      } else if (videoFileTypes.includes(path.extname(originalFilePath))) {
-        // Convert to audio (.wav) with ffmpeg
-        const audioFilePath = await convertVideoToAudio(originalFilePath, '.wav');
+      // TODO: #43 Reuse file the (converted) text file from the previous call -  temporary solution
+      const textFilePath = path.join(docFolder, file.originalname + '.txt');
 
-        // Convert audio to text
-        outputText = await convertSpeechToText(audioFilePath);
+      if (fs.existsSync(textFilePath)) {
+        outputText = (await readFile(textFilePath)).toString();
 
-        await fs.unlink(audioFilePath, () => {
-          console.log(`Converted audio file ${audioFilePath} deleted`);
-        });
+        console.log('CONVERTED TEXT OUTPUT ALREADY FOUND: ', textFilePath);
+      } else {
+        if (path.extname(originalFilePath) === '.docx') {
+          // Read docx file and get the plain text
+          outputText = await convertWordDocToText(originalFilePath);
+        } else if (viewerFileTypes.includes(path.extname(originalFilePath))) {
+          // Read the handwritten or any meaningful text directly
+          outputText = await convertHandwrittenFileToTextByAzure(originalFilePath);
+        } else if (audioFileTypes.includes(path.extname(originalFilePath))) {
+          console.log('AUDIO BEING PROCESSED: ', originalFilePath);
+          // Convert audio to text
+          outputText = await convertSpeechToText(originalFilePath);
+
+          console.log('AUDIO TEXT OUTPUT: ', outputText);
+        } else if (videoFileTypes.includes(path.extname(originalFilePath))) {
+          // Convert to audio (.wav) with ffmpeg
+          const audioFilePath = await convertVideoToAudio(originalFilePath, '.wav');
+
+          // Convert audio to text
+          outputText = await convertSpeechToText(audioFilePath);
+
+          await fs.unlink(audioFilePath, () => {
+            console.log(`Converted audio file ${audioFilePath} deleted`);
+          });
+        }
+
+        // Save text output file
+        await writeFile(textFilePath, outputText);
+
+        console.log('CONVERSION TEXT OUTPUT NOT FOUND, SAVING CONVERTED TEXT: ', textFilePath);
       }
 
       console.log('File content converted to text');
